refactor: cleanup welcomecontroller and rename to more fitting logincontroller
